ai-visual-testing/requirements.txt
playwright==1.48.0
openai==1.54.0
anthropic==0.40.0
google-generativeai==0.7.2
pillow==10.4.0
pyyaml==6.0.2
jinja2==3.1.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
typer==0.12.5
colorama==0.4.6
tenacity==9.0.0

